requirements.txt
fastapi
uvicorn[standard]
requests
pydantic-settings
python-dotenv
huggingface-hub
httpx
